Started better resizing logic, refactored

+++ src/type_utils.py
@@ -7,9 +7,9 @@ from typing import Optional
 
 ColorType = tuple[int, ...]
 
-BlitSequence = list[tuple[Surface, tuple[float, float]]]
+BlitSequence = list[tuple[Surface, tuple[int, int]]]
 # Last element is the layer
-LayeredBlitInfo = tuple[Surface, tuple[float, float], int]
+LayeredBlitInfo = tuple[Surface, tuple[int, int], int]
 LayeredBlitSequence = list[LayeredBlitInfo]
 # Contains the name, layer (can be None) and depth counter (used for nicer printing)
 LayerSequence = list[tuple[str, Optional[int], int]]
